+++ app/about/container.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import classNames from 'classnames/bind';
-import styles from 'css/components/about';
+import styles from './style';
 
 const cx = classNames.bind(styles);
