+++ app/reducers.js
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux';
-import user from 'reducers/user';
-import topic from 'reducers/topic';
-import message from 'reducers/message';
+import user from 'users/reducer';
+import topic from 'topics/reducer';
+import message from 'messages/reducer';
 import { routerReducer as routing } from 'react-router-redux';
 
 // Combine reducers with routeReducer which keeps track of
