+++ app/routes.jsx
@@ -3,7 +3,7 @@ import { Route, IndexRoute } from 'react-router';
 
 import App from 'containers/App';
 import Vote from 'topics';
-import About from 'containers/About';
+import About from 'about';
 import LoginOrRegister from 'containers/LoginOrRegister';
 import Dashboard from 'containers/Dashboard';
