+++ app/topics/reducer.js
